// packages/expo-router/src/native-tabs/NativeBottomTabs/types.ts
export type NativeTabOptions = Omit<
  BottomTabsScreenProps,
  | 'children'
  | 'placeholder'
  | 'onWillAppear'
  | 'onDidAppear'
  | 'onWillDisappear'
  | 'onDidDisappear'
  | 'isFocused'
  | 'tabKey'
  | 'icon'
  | 'selectedIcon'
  | 'iconResourceName'
  | 'specialEffects'
> &
  DefaultRouterOptions & {
    /**
     * The icon to display in the tab bar.
     * @platform android
     * @platform iOS
     */
    icon?: SfSymbolOrImageSource & {
      /**
       * The name of the drawable resource to use as an icon.
       * @platform android
       */
      drawable?: string;
    };
    /**
     * The icon to display when the tab is selected.
     * @platform iOS
     */
    selectedIcon?: SfSymbolOrImageSource;
  };

export type SfSymbolOrImageSource =
  | {
      /**
       * The name of the SF Symbol to use as an icon.
       * @platform iOS
       */
      sf?: SFSymbol;
    }
  | {
      /**
       * The image source to use as an icon.
       * @platform iOS
       */
      src?: ImageSourcePropType;
    };

export interface ExtendedNativeTabOptions extends NativeTabOptions {
  /**
   * If true, the tab will be hidden from the tab bar.
   */
  hidden?: boolean;
  specialEffects?: BottomTabsScreenProps['specialEffects'];
}

export interface NativeTabsStyleType {
  fontFamily?: TextStyle['fontFamily'];
  fontSize?: TextStyle['fontSize'];
  fontWeight?: TextStyle['fontWeight'];
  fontStyle?: TextStyle['fontStyle'];
  color?: TextStyle['color'];
  /**
   * @platform android
   * @platform iOS
   */
  iconColor?: ColorValue;
  backgroundColor?: ColorValue;
  /**
   * @platform iOS
   */
  blurEffect?: BottomTabsScreenProps['tabBarBlurEffect'];
  tintColor?: ColorValue;
  badgeBackgroundColor?: ColorValue;
  /**
   * @platform web
   */
  badgeTextColor?: ColorValue;
  /**
   * @platform android
   */
  rippleColor?: ColorValue;
  /**
   * @platform android
   */
  labelVisibilityMode?: TabBarItemLabelVisibilityMode;
  /**
   * @platform android
   * @platform web
   */
  '&:active'?: NativeTabsActiveStyleType;
}

export interface NativeTabsActiveStyleType {
  /**
   * @platform android
   * @platform web
   */
  color?: ColorValue;
  /**
   * @platform android
   * @platform web
   */
  fontSize?: TextStyle['fontSize'];
  /**
   * @platform android
   */
  iconColor?: ColorValue;
  /**
   * @platform android
   * @platform web
   */
  indicatorColor?: ColorValue;
}

export interface NativeTabsProps extends PropsWithChildren {
  style?: NativeTabsStyleType;
  /**
   * https://developer.apple.com/documentation/uikit/uitabbarcontroller/tabbarminimizebehavior
   *
   * Supported values:
   * - `none` - The tab bar does not minimize.
   * - `onScrollUp` - The tab bar minimizes when scrolling up, and expands when scrolling back down. Recommended if the scroll view content is aligned to the bottom.
   * - `onScrollDown` - The tab bar minimizes when scrolling down, and expands when scrolling back up.
   * - `automatic` - Resolves to the system default minimize behavior.
   *
   * @default automatic
   *
   * @platform iOS 26
   */
  minimizeBehavior?: BottomTabsProps['tabBarMinimizeBehavior'];
  /**
   * Disables the active indicator for the tab bar.
   *
   * @platform android
   */
  disableIndicator?: boolean;
  /**
   * The behavior when navigating back with the back button.
   *
   * @platform android
   */
  backBehavior?: 'none' | 'initialRoute' | 'history';
}

export interface NativeTabsViewProps extends NativeTabsProps {
  builder: ReturnType<
    typeof useNavigationBuilder<
      TabNavigationState<ParamListBase>,
      TabRouterOptions,
      Record<string, (...args: any) => void>,
      NativeTabOptions,
      Record<string, any>
    >
  >;
}